local constants
