+++ components/Table/TableRowColumn.jsx
@@ -4,6 +4,6 @@ import { getWidthStyle } from './tableHelpers';
 
 export default ({ row, column, totalWidth }) => (
   <MaterialTableRowColumn style={getWidthStyle(column.width, totalWidth)}>
-    {row && column.dispFn ? column.dispFn(row, column) : (row[column.field+"Display"] || row[column.field])}
+    {row && column.dispFn ? column.dispFn(row, column, row[column.field]) : row[column.field]}
   </MaterialTableRowColumn>
 );
